requirements.txt
python-telegram-bot==21.10
flask==3.1.0
gunicorn==23.0.0